Fix compiler error
Fixes compiler error: The package javax.xml.<artifactId> is accessible from more than one module.
Currently this is not showing in many javac comiplers due to a bug in the compiler: https://bugs.openjdk.java.net/browse/JDK-8215739
Once this bug is fixed and the error is reported correctly this will not build correctly. Excluding xml-apis:xml-apis from the classpath will fix this.