ACS-3476: Re-adding a getAdditional method
It was needed for proper processing of ConfigurationProperties annotation and allowing for TRANSFORMER_ROUTES_ADDITIONAL_<engineName> for backwards compatibility.
Spring mechanism takes the "transform.routes" gets the "additional" as a setting and adds to map ("ai", configFile).
Without this get, the spring mechanism can't add this property due to no access to additional variable